Make gruvbox show right color.
This patch makes a side effect which is not intended.

For a certain shell, after the first quitting vim, multiple blank lines
show up.

This might be a vim bug or something.

When executing the line causing the problem through the vim command
line, there's no problem. It occurs only if the line is executed through
the vim script.